requirements.txt
smartsheet-python-sdk
requests
certifi
# optional: add aws-secretsmanager or boto3 later when moving to AWS Lambda
